refactor of the state transition functions
