+++ lib/lambda_ethereum_consensus/state_transition/epoch_processing.ex
@@ -4,6 +4,7 @@ defmodule LambdaEthereumConsensus.StateTransition.EpochProcessing do
   """
 
   alias LambdaEthereumConsensus.StateTransition.Accessors
+  alias LambdaEthereumConsensus.StateTransition.Misc
   alias LambdaEthereumConsensus.StateTransition.Predicates
   alias SszTypes.BeaconState
   alias SszTypes.Validator
@@ -115,21 +116,21 @@ defmodule LambdaEthereumConsensus.StateTransition.EpochProcessing do
           Accessors.get_previous_epoch(state)
         )
 
-      state_is_in_inactivity_leak = is_in_inactivity_leak(state)
+      state_is_in_inactivity_leak = Predicates.is_in_inactivity_leak(state)
 
       updated_eligible_validator_indices =
-        get_eligible_validator_indices(state)
+        Accessors.get_eligible_validator_indices(state)
         |> Enum.map(fn index ->
           inactivity_score = Enum.at(state.inactivity_scores, index)
 
           new_inactivity_score =
-            increase_inactivity_score(
+            Misc.increase_inactivity_score(
               inactivity_score,
               index,
               unslashed_participating_indices,
               inactivity_score_bias
             )
-            |> decrease_inactivity_score(
+            |> Misc.decrease_inactivity_score(
               state_is_in_inactivity_leak,
               inactivity_score_recovery_rate
             )
@@ -142,75 +143,15 @@ defmodule LambdaEthereumConsensus.StateTransition.EpochProcessing do
         state.inactivity_scores
         |> Stream.with_index()
         |> Stream.map(fn {inactivity_score, index} ->
-          update_inactivity_score(updated_eligible_validator_indices, index, inactivity_score)
+          Misc.update_inactivity_score(
+            updated_eligible_validator_indices,
+            index,
+            inactivity_score
+          )
         end)
         |> Enum.to_list()
 
       {:ok, %{state | inactivity_scores: updated_inactive_scores}}
     end
-  end
-
-  @spec increase_inactivity_score(SszTypes.uint64(), integer, MapSet.t(), SszTypes.uint64()) ::
-          SszTypes.uint64()
-  defp increase_inactivity_score(
-         inactivity_score,
-         index,
-         unslashed_participating_indices,
-         inactivity_score_bias
-       ) do
-    if MapSet.member?(unslashed_participating_indices, index) do
-      inactivity_score - min(1, inactivity_score)
-    else
-      inactivity_score + inactivity_score_bias
-    end
-  end
-
-  @spec decrease_inactivity_score(SszTypes.uint64(), boolean, SszTypes.uint64()) ::
-          SszTypes.uint64()
-  defp decrease_inactivity_score(
-         inactivity_score,
-         state_is_in_inactivity_leak,
-         inactivity_score_recovery_rate
-       ) do
-    if state_is_in_inactivity_leak do
-      inactivity_score
-    else
-      inactivity_score - min(inactivity_score_recovery_rate, inactivity_score)
-    end
-  end
-
-  @spec update_inactivity_score(%{integer => SszTypes.uint64()}, integer, {SszTypes.uint64()}) ::
-          SszTypes.uint64()
-  defp update_inactivity_score(updated_eligible_validator_indices, index, inactivity_score) do
-    case Map.fetch(updated_eligible_validator_indices, index) do
-      {:ok, new_eligible_validator_index} -> new_eligible_validator_index
-      :error -> inactivity_score
-    end
-  end
-
-  @spec is_in_inactivity_leak(BeaconState.t()) :: boolean
-  def is_in_inactivity_leak(%BeaconState{} = state) do
-    min_epochs_to_inactivity_penalty = ChainSpec.get("MIN_EPOCHS_TO_INACTIVITY_PENALTY")
-
-    get_finality_delay(state) > min_epochs_to_inactivity_penalty
-  end
-
-  @spec get_finality_delay(BeaconState.t()) :: SszTypes.uint64()
-  def get_finality_delay(%BeaconState{} = state) do
-    Accessors.get_previous_epoch(state) - state.finalized_checkpoint.epoch
-  end
-
-  @spec get_eligible_validator_indices(BeaconState.t()) :: list(SszTypes.validator_index())
-  def get_eligible_validator_indices(%BeaconState{validators: validators} = state) do
-    previous_epoch = Accessors.get_previous_epoch(state)
-
-    validators
-    |> Stream.with_index()
-    |> Stream.filter(fn {validator, _index} ->
-      Predicates.is_active_validator(validator, previous_epoch) ||
-        (validator.slashed && previous_epoch + 1 < validator.withdrawable_epoch)
-    end)
-    |> Stream.map(fn {_validator, index} -> index end)
-    |> Enum.to_list()
   end
 end
