+++ lib/lambda_ethereum_consensus/state_transition/misc.ex
@@ -11,4 +11,42 @@ defmodule LambdaEthereumConsensus.StateTransition.Misc do
     slots_per_epoch = ChainSpec.get("SLOTS_PER_EPOCH")
     div(slot, slots_per_epoch)
   end
+
+  @spec increase_inactivity_score(SszTypes.uint64(), integer, MapSet.t(), SszTypes.uint64()) ::
+          SszTypes.uint64()
+  def increase_inactivity_score(
+        inactivity_score,
+        index,
+        unslashed_participating_indices,
+        inactivity_score_bias
+      ) do
+    if MapSet.member?(unslashed_participating_indices, index) do
+      inactivity_score - min(1, inactivity_score)
+    else
+      inactivity_score + inactivity_score_bias
+    end
+  end
+
+  @spec decrease_inactivity_score(SszTypes.uint64(), boolean, SszTypes.uint64()) ::
+          SszTypes.uint64()
+  def decrease_inactivity_score(
+        inactivity_score,
+        state_is_in_inactivity_leak,
+        inactivity_score_recovery_rate
+      ) do
+    if state_is_in_inactivity_leak do
+      inactivity_score
+    else
+      inactivity_score - min(inactivity_score_recovery_rate, inactivity_score)
+    end
+  end
+
+  @spec update_inactivity_score(%{integer => SszTypes.uint64()}, integer, {SszTypes.uint64()}) ::
+          SszTypes.uint64()
+  def update_inactivity_score(updated_eligible_validator_indices, index, inactivity_score) do
+    case Map.fetch(updated_eligible_validator_indices, index) do
+      {:ok, new_eligible_validator_index} -> new_eligible_validator_index
+      :error -> inactivity_score
+    end
+  end
 end
